+++ tests/Unit/Geometry/ResizerTest.php
@@ -7,48 +7,48 @@ namespace Intervention\Image\Tests\Unit\Geometry;
 use Intervention\Image\Alignment;
 use PHPUnit\Framework\Attributes\CoversClass;
 use Intervention\Image\Geometry\Rectangle;
-use Intervention\Image\Geometry\Tools\RectangleResizer;
+use Intervention\Image\Geometry\Tools\Resizer;
 use Intervention\Image\Tests\Providers\ResizeDataProvider;
 use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\TestCase;
 
-#[CoversClass(RectangleResizer::class)]
-final class RectangleResizerTest extends TestCase
+#[CoversClass(Resizer::class)]
+final class ResizerTest extends TestCase
 {
     public function testMake(): void
     {
-        $resizer = RectangleResizer::to();
-        $this->assertInstanceOf(RectangleResizer::class, $resizer);
+        $resizer = Resizer::to();
+        $this->assertInstanceOf(Resizer::class, $resizer);
 
-        $resizer = RectangleResizer::to(height: 100);
-        $this->assertInstanceOf(RectangleResizer::class, $resizer);
+        $resizer = Resizer::to(height: 100);
+        $this->assertInstanceOf(Resizer::class, $resizer);
 
-        $resizer = RectangleResizer::to(100);
-        $this->assertInstanceOf(RectangleResizer::class, $resizer);
+        $resizer = Resizer::to(100);
+        $this->assertInstanceOf(Resizer::class, $resizer);
 
-        $resizer = RectangleResizer::to(100, 100);
-        $this->assertInstanceOf(RectangleResizer::class, $resizer);
+        $resizer = Resizer::to(100, 100);
+        $this->assertInstanceOf(Resizer::class, $resizer);
     }
 
     public function testToWidth(): void
     {
-        $resizer = new RectangleResizer();
+        $resizer = new Resizer();
         $result = $resizer->toWidth(100);
-        $this->assertInstanceOf(RectangleResizer::class, $result);
+        $this->assertInstanceOf(Resizer::class, $result);
     }
 
     public function testToHeight(): void
     {
-        $resizer = new RectangleResizer();
+        $resizer = new Resizer();
         $result = $resizer->toHeight(100);
-        $this->assertInstanceOf(RectangleResizer::class, $result);
+        $this->assertInstanceOf(Resizer::class, $result);
     }
 
     public function testToSize(): void
     {
-        $resizer = new RectangleResizer();
+        $resizer = new Resizer();
         $resizer = $resizer->toSize(new Rectangle(200, 100));
-        $this->assertInstanceOf(RectangleResizer::class, $resizer);
+        $this->assertInstanceOf(Resizer::class, $resizer);
     }
 
     /**
@@ -57,7 +57,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'resizeDataProvider')]
     public function testResize(Rectangle $input, array $resizeParameters, Rectangle $result): void
     {
-        $resizer = new RectangleResizer(...$resizeParameters);
+        $resizer = new Resizer(...$resizeParameters);
         $resized = $resizer->resize($input);
         $this->assertEquals($result->width(), $resized->width());
         $this->assertEquals($result->height(), $resized->height());
@@ -69,7 +69,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'resizeDownDataProvider')]
     public function testResizeDown(Rectangle $input, array $resizeParameters, Rectangle $result): void
     {
-        $resizer = new RectangleResizer(...$resizeParameters);
+        $resizer = new Resizer(...$resizeParameters);
         $resized = $resizer->resizeDown($input);
         $this->assertEquals($result->width(), $resized->width());
         $this->assertEquals($result->height(), $resized->height());
@@ -81,7 +81,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'scaleDataProvider')]
     public function testScale(Rectangle $input, array $resizeParameters, Rectangle $result): void
     {
-        $resizer = new RectangleResizer(...$resizeParameters);
+        $resizer = new Resizer(...$resizeParameters);
         $resized = $resizer->scale($input);
         $this->assertEquals($result->width(), $resized->width());
         $this->assertEquals($result->height(), $resized->height());
@@ -93,7 +93,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'scaleDownDataProvider')]
     public function testScaleDown(Rectangle $input, array $resizeParameters, Rectangle $result): void
     {
-        $resizer = new RectangleResizer(...$resizeParameters);
+        $resizer = new Resizer(...$resizeParameters);
         $resized = $resizer->scaleDown($input);
         $this->assertEquals($result->width(), $resized->width());
         $this->assertEquals($result->height(), $resized->height());
@@ -102,7 +102,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'coverDataProvider')]
     public function testCover(Rectangle $origin, Rectangle $target, Rectangle $result): void
     {
-        $resizer = new RectangleResizer();
+        $resizer = new Resizer();
         $resizer->toSize($target);
         $resized = $resizer->cover($origin);
         $this->assertEquals($result->width(), $resized->width());
@@ -112,7 +112,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'containDataProvider')]
     public function testContain(Rectangle $origin, Rectangle $target, Rectangle $result): void
     {
-        $resizer = new RectangleResizer();
+        $resizer = new Resizer();
         $resizer->toSize($target);
         $resized = $resizer->contain($origin);
         $this->assertEquals($result->width(), $resized->width());
@@ -122,7 +122,7 @@ final class RectangleResizerTest extends TestCase
     #[DataProviderExternal(ResizeDataProvider::class, 'cropDataProvider')]
     public function testCrop(Rectangle $origin, Rectangle $target, string|Alignment $position, Rectangle $result): void
     {
-        $resizer = new RectangleResizer();
+        $resizer = new Resizer();
         $resizer->toSize($target);
         $resized = $resizer->crop($origin, $position);
         $this->assertEquals($result->width(), $resized->width());